VIRO-889,VIRO-890: LifeCycle Crashes for GVR and OVR
- Both gvr and ovr layouts were listening for lifecycle callbacks that
may not have originated from the React activity, causing onPause to
be called on activities that have already paused. Thus fix is to
save a weak reference of the react activity to compare against.

- Fixed Memory-Leak-OVR-LifeCycle acyclic dependency issue where Viro Controls
that had references to the renderer were not being deallocated properly. This
is because even though they were queued to be deallocted on the RenderThread,
the onDestroy Msg required by OVRRenderer to process that queue is only processed
on the destruction of OVRRenderer, of which it's ref is held by the very Viro
Controls that cannot be destroyed. Thus, fix is to trigger an onDestory Msg apart
from the destructor to break that dependency, allowing for proper deallocations
to occur.

+++ android/app/src/main/java/com/viro/renderer/jni/ViroGvrLayout.java
@@ -24,6 +24,7 @@ import com.google.vr.ndk.base.GvrLayout;
 import com.viro.renderer.FrameListener;
 import com.viro.renderer.GLSurfaceViewQueue;
 
+import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 import java.util.List;
 import javax.microedition.khronos.egl.EGLConfig;
@@ -48,6 +49,7 @@ public class ViroGvrLayout extends GvrLayout implements VrView {
     private List<FrameListener> mFrameListeners = new ArrayList();
     private GlListener mGlListener = null;
     private PlatformUtil mPlatformUtil;
+    private WeakReference<Activity> mWeakActivity;
 
     public ViroGvrLayout(Context context, GlListener glListener) {
         super(context);
@@ -138,6 +140,7 @@ public class ViroGvrLayout extends GvrLayout implements VrView {
         // Prevent screen from switching to portrait
         ((Activity)activityContext).setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE);
 
+        mWeakActivity = new WeakReference<Activity>((Activity)getContext());
         Application app = (Application)activityContext.getApplicationContext();
         app.registerActivityLifecycleCallbacks(this);
     }
@@ -200,12 +203,20 @@ public class ViroGvrLayout extends GvrLayout implements VrView {
 
     @Override
     public void onActivityPaused(Activity activity) {
+        if (mWeakActivity.get() != activity){
+            return;
+        }
+
         mNativeRenderer.onPause();
         super.onPause();
     }
 
     @Override
     public void onActivityResumed(Activity activity) {
+        if (mWeakActivity.get() != activity){
+            return;
+        }
+
         mNativeRenderer.onResume();
 
         // Ensure fullscreen immersion.
@@ -228,9 +239,16 @@ public class ViroGvrLayout extends GvrLayout implements VrView {
 
     @Override
     public void onActivityDestroyed(Activity activity) {
+        if (mWeakActivity.get() != activity){
+            return;
+        }
+
         super.shutdown();
         mNativeRenderContext.delete();
         mNativeRenderer.destroy();
+
+        Application app = (Application)activity.getApplicationContext();
+        app.unregisterActivityLifecycleCallbacks(this);
     }
 
     @Override
@@ -263,11 +281,19 @@ public class ViroGvrLayout extends GvrLayout implements VrView {
 
     @Override
     public void onActivityStarted(Activity activity) {
+        if (mWeakActivity.get() != activity){
+            return;
+        }
+
         mNativeRenderer.onStart();
     }
 
     @Override
     public void onActivityStopped(Activity activity) {
+        if (mWeakActivity.get() != activity){
+            return;
+        }
+
         mNativeRenderer.onStop();
     }
